+++ modules/Material/PopupBase.qml
@@ -24,6 +24,7 @@ FocusScope {
 
     property color overlayColor: "transparent"
     property string overlayLayer: "overlayLayer"
+    property bool globalMouseAreaEnabled: true
 
     property bool showing: false