+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 QML Material CHANGELOG
 ======================
 
+ * Add a ProgressCircle component following the Google spec
+
 ### Version 0.0.5 (March 4, 2015)
 
  * Move popovers as necessary to prevent them being clipped by the window