+++ CHANGELOG.md
@@ -2,6 +2,7 @@ QML Material CHANGELOG
 ======================
 
  * Add a ProgressCircle component following the Google spec
+ * Add a tooltip component and use it in the ActionBar
 
 ### Version 0.0.5 (March 4, 2015)