+++ CHANGELOG.md
@@ -1,11 +1,21 @@
 QML Material CHANGELOG
 ======================
 
+### Version 0.0.6 (April 16, 2015)
+
  * Add a ProgressCircle component following the Google spec
  * Add a tooltip component and use it in the ActionBar
- * [Tooltip] Clean up the code and improve the style
- * [Tooltip] Clean up the code and implement desktop-sized tooltips
  * Add an OverlayView component for view transitions  
+ * [Snackbar] Bug fixes and style improvements
+ * [Button] Remove default Qt Quick Controls padding
+ * [SimpleMenu/MenuField] Fixes some binding loops
+ * [TextField] Fixes floating label behavior on devices
+ * [Action] Support for keybindings in Action
+ * [ListItem] Fix bug with Subtitled action not showing
+ * [ListItem] Add support for a custom view in the subtitled list item
+ * [Button] Add disabled state to button
+ * General bug fixing and code refactoring
+
 
 ### Version 0.0.5 (March 4, 2015)